Dialog, CampaignBoxをコンポーネントへ分割
